m2p2_v2/frontend/requirements.txt
requests
streamlit
typing_extensions
pandas
scikit-learn == 1.0.2
tensorflow-cpu == 2.8.2
numpy
Pillow